Adapt FileDownload to show events on calendar
